+++ src/ChangeLog.php
@@ -3,7 +3,7 @@
  * PHP Version 5.5
  * @category Library
  * @package ChangeLog
- * @author Steve "uru" West <steve.david.west@gmail.com>
+ * @author Steve "uru" West <steven.david.west@gmail.com>
  * @license MIT http://opensource.org/licenses/MIT
  * @link https://github.com/stevewest/changelog
  */