+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## [0.4.0] - 2015-01-29
 ### Added
 - Remove a `Release` from a `Log` [#14](https://github.com/stevewest/changelog/issues/14).
 - Added date field to `Release` and date parsing to `KeepAChangeLog` parser [#11](https://github.com/stevewest/changelog/issues/11).
@@ -38,6 +40,7 @@
 - Initial version.
 
 [Unreleased]: https://github.com/stevewest/changelog
+[0.4.0]: https://github.com/stevewest/changelog/releases/tag/0.4.0
 [0.3.0]: https://github.com/stevewest/changelog/releases/tag/0.3.0
 [0.2.0]: https://github.com/stevewest/changelog/releases/tag/0.2.0
 [0.1.0]: https://github.com/stevewest/changelog/releases/tag/0.1.0